core: move enums to folder
