+++ BlasII.Randomizer.MapTracker/Colors.cs
@@ -1,20 +1,16 @@
 ﻿using System.Collections.Generic;
+using BlasII.Randomizer.MapTracker.Enums;
 using UnityEngine;
 
 namespace BlasII.Randomizer.MapTracker;
 
 internal static class Colors
 {
-    public static readonly Dictionary<Logic, Color> LogicColors = new()
+    public static Dictionary<Logic, Color> LogicColors { get; } = new()
     {
-        { Logic.Finished, RGBColor(63, 63, 63) },
-        { Logic.NoneReachable, RGBColor(207, 16, 16) },
-        { Logic.SomeReachable, RGBColor(255, 159, 32) },
-        { Logic.AllReachable, RGBColor(32, 255, 32) },
+        { Logic.Finished, new Color32(63, 63, 63, 1) },
+        { Logic.NoneReachable, new Color32(207, 16, 16, 1) },
+        { Logic.SomeReachable, new Color32(255, 159, 32, 1) },
+        { Logic.AllReachable, new Color32(32, 255, 32, 1) },
     };
-
-    private static Color RGBColor(int r, int g, int b)
-    {
-        return new Color(r / 255f, g / 255f, b / 255f);
-    }
 }
